Convert create-react-class usage with ES2015 class
