+++ .babelrc
@@ -1,7 +1,8 @@
 {
   "presets":[
     "react",
-    ["es2015", {"loose": true}]
+    ["es2015", {"loose": true}],
+    "stage-2"
   ],
   "plugins": ["transform-object-assign", "transform-object-rest-spread"]
 }
