+++ package.json
@@ -36,6 +36,7 @@
     "babel-preset-airbnb": "^2.1.1",
     "babel-preset-es2015": "^6.16.0",
     "babel-preset-react": "^6.16.0",
+    "babel-preset-stage-2": "^6.24.1",
     "css-loader": "^0.28.0",
     "deepmerge": "^1.1.0",
     "del": "^2.2.2",
@@ -66,7 +67,6 @@
   "dependencies": {
     "can-use-dom": "^0.1.0",
     "classnames": "^2.2.5",
-    "create-react-class": "^15.5.2",
     "json2mq": "^0.2.0",
     "object-assign": "^4.1.0",
     "slick-carousel": "^1.6.0",
