+++ src/inner-slider.js
@@ -1,37 +1,41 @@
 'use strict';
 
 import React from 'react';
-import EventHandlersMixin from './mixins/event-handlers';
-import HelpersMixin from './mixins/helpers';
+import ReactDOM from 'react-dom';
 import initialState from './initial-state';
 import defaultProps from './default-props';
-import createReactClass from 'create-react-class';
 import classnames from 'classnames';
 import assign from 'object-assign';
 
+import canGoNext from './helpers/canGoNext';
+import {getTrackCSS, getTrackLeft, getTrackAnimateCSS} from './helpers/trackHelper';
 import {Track} from './track';
 import {Dots} from './dots';
 import {PrevArrow, NextArrow} from './arrows';
 
-export var InnerSlider = createReactClass({
-  mixins: [HelpersMixin, EventHandlersMixin],
-  list: null,
-  track: null,
-  listRefHandler: function (ref) {
-    this.list = ref;
-  },
-  trackRefHandler: function (ref) {
-    this.track = ref;
-  },
-  getInitialState: function () {
-    return Object.assign({}, initialState, {
-      currentSlide: this.props.initialSlide
+export class InnerSlider extends React.Component {
+  static defaultProps = defaultProps;
+
+  list = null;
+  track = null;
+
+  constructor(props, context) {
+    super(props, context);
+
+    this.state = Object.assign({}, initialState, {
+      currentSlide: props.initialSlide
     });
-  },
-  getDefaultProps: function () {
-    return defaultProps;
-  },
-  componentWillMount: function () {
+  }
+
+  listRefHandler = (ref) => {
+    this.list = ref;
+  };
+
+  trackRefHandler = (ref) => {
+    this.track = ref;
+  };
+
+  componentWillMount() {
     if (this.props.init) {
       this.props.init();
     }
@@ -50,8 +54,9 @@ export var InnerSlider = createReactClass({
         lazyLoadedList: lazyLoadedList
       });
     }
-  },
-  componentDidMount: function componentDidMount() {
+  }
+
+  componentDidMount() {
     // Hack for autoplay -- Inspect Later
     this.initialize(this.props);
     this.adaptHeight();
@@ -65,8 +70,9 @@ export var InnerSlider = createReactClass({
     } else {
       window.attachEvent('onresize', this.onWindowResized);
     }
-  },
-  componentWillUnmount: function componentWillUnmount() {
+  }
+
+  componentWillUnmount() {
     if (this.animationEndCallback) {
       clearTimeout(this.animationEndCallback);
     }
@@ -78,8 +84,9 @@ export var InnerSlider = createReactClass({
     if (this.state.autoPlayTimer) {
       clearInterval(this.state.autoPlayTimer);
     }
-  },
-  componentWillReceiveProps: function(nextProps) {
+  }
+
+  componentWillReceiveProps(nextProps) {
     if (this.props.slickGoTo != nextProps.slickGoTo) {
       if (process.env.NODE_ENV !== 'production') {
         console.warn('react-slick deprecation warning: slickGoTo prop is deprecated and it will be removed in next release. Use slickGoTo method instead')
@@ -99,11 +106,13 @@ export var InnerSlider = createReactClass({
     } else {
       this.update(nextProps);
     }
-  },
-  componentDidUpdate: function () {
+  }
+
+  componentDidUpdate() {
     this.adaptHeight();
-  },
-  onWindowResized: function () {
+  }
+
+  onWindowResized = () => {
     this.update(this.props);
     // animating state should be cleared while resizing, otherwise autoplay stops working
     this.setState({
@@ -111,21 +120,25 @@ export var InnerSlider = createReactClass({
     });
     clearTimeout(this.animationEndCallback);
     delete this.animationEndCallback;
-  },
-  slickPrev: function () {
+  };
+
+  slickPrev = () => {
     this.changeSlide({message: 'previous'});
-  },
-  slickNext: function () {
+  };
+
+  slickNext = () => {
     this.changeSlide({message: 'next'});
-  },
-  slickGoTo: function (slide) {
+  };
+
+  slickGoTo = (slide) => {
     typeof slide === 'number' && this.changeSlide({
       message: 'index',
       index: slide,
       currentSlide: this.state.currentSlide
     });
-  },
-  render: function () {
+  };
+
+  render() {
     var className = classnames('slick-initialized', 'slick-slider', this.props.className, {
       'slick-vertical': this.props.vertical,
     });
@@ -240,4 +253,660 @@ export var InnerSlider = createReactClass({
       </div>
     );
   }
-});
+
+  initialize = (props) => {
+    const slickList = ReactDOM.findDOMNode(this.list);
+
+    var slideCount = React.Children.count(props.children);
+    var listWidth = this.getWidth(slickList);
+    var trackWidth = this.getWidth(ReactDOM.findDOMNode(this.track));
+    var slideWidth;
+
+    if (!props.vertical) {
+      var centerPaddingAdj = props.centerMode && (parseInt(props.centerPadding) * 2);
+      slideWidth = (this.getWidth(ReactDOM.findDOMNode(this)) - centerPaddingAdj)/props.slidesToShow;
+    } else {
+      slideWidth = this.getWidth(ReactDOM.findDOMNode(this));
+    }
+
+    const slideHeight = this.getHeight(slickList.querySelector('[data-index="0"]'));
+    const listHeight = slideHeight * props.slidesToShow;
+
+    var currentSlide = props.rtl ? slideCount - 1 - props.initialSlide : props.initialSlide;
+
+    this.setState({
+      slideCount,
+      slideWidth,
+      listWidth,
+      trackWidth,
+      currentSlide,
+      slideHeight,
+      listHeight,
+    }, function () {
+
+      var targetLeft = getTrackLeft(assign({
+        slideIndex: this.state.currentSlide,
+        trackRef: this.track
+      }, props, this.state));
+      // getCSS function needs previously set state
+      var trackStyle = getTrackCSS(assign({left: targetLeft}, props, this.state));
+
+      this.setState({trackStyle: trackStyle});
+
+      this.autoPlay(); // once we're set up, trigger the initial autoplay.
+    });
+  };
+
+  update = (props) => {
+    const slickList = ReactDOM.findDOMNode(this.list);
+    // This method has mostly same code as initialize method.
+    // Refactor it
+    var slideCount = React.Children.count(props.children);
+    var listWidth = this.getWidth(slickList);
+    var trackWidth = this.getWidth(ReactDOM.findDOMNode(this.track));
+    var slideWidth;
+
+    if (!props.vertical) {
+      var centerPaddingAdj = props.centerMode && (parseInt(props.centerPadding) * 2);
+      slideWidth = (this.getWidth(ReactDOM.findDOMNode(this)) - centerPaddingAdj)/props.slidesToShow;
+    } else {
+      slideWidth = this.getWidth(ReactDOM.findDOMNode(this));
+    }
+
+    const slideHeight = this.getHeight(slickList.querySelector('[data-index="0"]'));
+    const listHeight = slideHeight * props.slidesToShow;
+
+    // pause slider if autoplay is set to false
+    if(props.autoplay) {
+      this.pause();
+    } else {
+      this.autoPlay();
+    }
+
+    this.setState({
+      slideCount,
+      slideWidth,
+      listWidth,
+      trackWidth,
+      slideHeight,
+      listHeight,
+    }, function () {
+
+      var targetLeft = getTrackLeft(assign({
+        slideIndex: this.state.currentSlide,
+        trackRef: this.track
+      }, props, this.state));
+      // getCSS function needs previously set state
+      var trackStyle = getTrackCSS(assign({left: targetLeft}, props, this.state));
+
+      this.setState({trackStyle: trackStyle});
+    });
+  };
+
+  getWidth = (elem) => {
+    return elem.getBoundingClientRect().width || elem.offsetWidth || 0;
+  };
+
+  getHeight = (elem) => {
+    return elem.getBoundingClientRect().height || elem.offsetHeight || 0;
+  };
+
+  adaptHeight = () => {
+    if (this.props.adaptiveHeight) {
+      var selector = '[data-index="' + this.state.currentSlide +'"]';
+      if (this.list) {
+        var slickList = ReactDOM.findDOMNode(this.list);
+        slickList.style.height = slickList.querySelector(selector).offsetHeight + 'px';
+      }
+    }
+  };
+
+  slideHandler = (index) => {
+    // Functionality of animateSlide and postSlide is merged into this function
+    // console.log('slideHandler', index);
+    var targetSlide, currentSlide;
+    var targetLeft, currentLeft;
+    var callback;
+
+    if (this.props.waitForAnimate && this.state.animating) {
+      return;
+    }
+
+    if (this.props.fade) {
+      currentSlide = this.state.currentSlide;
+
+      // Don't change slide if it's not infite and current slide is the first or last slide.
+      if(this.props.infinite === false &&
+        (index < 0 || index >= this.state.slideCount)) {
+        return;
+      }
+
+      //  Shifting targetSlide back into the range
+      if (index < 0) {
+        targetSlide = index + this.state.slideCount;
+      } else if (index >= this.state.slideCount) {
+        targetSlide = index - this.state.slideCount;
+      } else {
+        targetSlide = index;
+      }
+
+      if (this.props.lazyLoad && this.state.lazyLoadedList.indexOf(targetSlide) < 0) {
+        this.setState({
+          lazyLoadedList: this.state.lazyLoadedList.concat(targetSlide)
+        });
+      }
+
+      callback = () => {
+        this.setState({
+          animating: false
+        });
+        if (this.props.afterChange) {
+          this.props.afterChange(targetSlide);
+        }
+        delete this.animationEndCallback;
+      };
+
+      this.setState({
+        animating: true,
+        currentSlide: targetSlide
+      }, function () {
+        this.animationEndCallback = setTimeout(callback, this.props.speed);
+      });
+
+      if (this.props.beforeChange) {
+        this.props.beforeChange(this.state.currentSlide, targetSlide);
+      }
+
+      this.autoPlay();
+      return;
+    }
+
+    targetSlide = index;
+    if (targetSlide < 0) {
+      if(this.props.infinite === false) {
+        currentSlide = 0;
+      } else if (this.state.slideCount % this.props.slidesToScroll !== 0) {
+        currentSlide = this.state.slideCount - (this.state.slideCount % this.props.slidesToScroll);
+      } else {
+        currentSlide = this.state.slideCount + targetSlide;
+      }
+    } else if (targetSlide >= this.state.slideCount) {
+      if(this.props.infinite === false) {
+        currentSlide = this.state.slideCount - this.props.slidesToShow;
+      } else if (this.state.slideCount % this.props.slidesToScroll !== 0) {
+        currentSlide = 0;
+      } else {
+        currentSlide = targetSlide - this.state.slideCount;
+      }
+    } else {
+      currentSlide = targetSlide;
+    }
+
+    targetLeft = getTrackLeft(assign({
+      slideIndex: targetSlide,
+      trackRef: this.track
+    }, this.props, this.state));
+
+    currentLeft = getTrackLeft(assign({
+      slideIndex: currentSlide,
+      trackRef: this.track
+    }, this.props, this.state));
+
+    if (this.props.infinite === false) {
+      targetLeft = currentLeft;
+    }
+
+    if (this.props.beforeChange) {
+      this.props.beforeChange(this.state.currentSlide, currentSlide);
+    }
+
+    if (this.props.lazyLoad) {
+      var loaded = true;
+      var slidesToLoad = [];
+      for (var i = targetSlide; i < targetSlide + this.props.slidesToShow; i++ ) {
+        loaded = loaded && (this.state.lazyLoadedList.indexOf(i) >= 0);
+        if (!loaded) {
+          slidesToLoad.push(i);
+        }
+      }
+      if (!loaded) {
+        this.setState({
+          lazyLoadedList: this.state.lazyLoadedList.concat(slidesToLoad)
+        });
+      }
+    }
+
+    // Slide Transition happens here.
+    // animated transition happens to target Slide and
+    // non - animated transition happens to current Slide
+    // If CSS transitions are false, directly go the current slide.
+
+    if (this.props.useCSS === false) {
+
+      this.setState({
+        currentSlide: currentSlide,
+        trackStyle: getTrackCSS(assign({left: currentLeft}, this.props, this.state))
+      }, function () {
+        if (this.props.afterChange) {
+          this.props.afterChange(currentSlide);
+        }
+      });
+
+    } else {
+
+      var nextStateChanges = {
+        animating: false,
+        currentSlide: currentSlide,
+        trackStyle: getTrackCSS(assign({left: currentLeft}, this.props, this.state)),
+        swipeLeft: null
+      };
+
+      callback = () => {
+        this.setState(nextStateChanges);
+        if (this.props.afterChange) {
+          this.props.afterChange(currentSlide);
+        }
+        delete this.animationEndCallback;
+      };
+
+      this.setState({
+        animating: true,
+        currentSlide: currentSlide,
+        trackStyle: getTrackAnimateCSS(assign({left: targetLeft}, this.props, this.state))
+      }, function () {
+        this.animationEndCallback = setTimeout(callback, this.props.speed);
+      });
+
+    }
+
+    this.autoPlay();
+  };
+
+  swipeDirection = (touchObject) => {
+    var xDist, yDist, r, swipeAngle;
+
+    xDist = touchObject.startX - touchObject.curX;
+    yDist = touchObject.startY - touchObject.curY;
+    r = Math.atan2(yDist, xDist);
+
+    swipeAngle = Math.round(r * 180 / Math.PI);
+    if (swipeAngle < 0) {
+        swipeAngle = 360 - Math.abs(swipeAngle);
+    }
+    if ((swipeAngle <= 45) && (swipeAngle >= 0) || (swipeAngle <= 360) && (swipeAngle >= 315)) {
+        return (this.props.rtl === false ? 'left' : 'right');
+    }
+    if ((swipeAngle >= 135) && (swipeAngle <= 225)) {
+        return (this.props.rtl === false ? 'right' : 'left');
+    }
+    if (this.props.verticalSwiping === true) {
+      if ((swipeAngle >= 35) && (swipeAngle <= 135)) {
+        return 'down';
+      } else {
+        return 'up';
+      }
+    }
+
+    return 'vertical';
+  };
+
+  play = () => {
+    var nextIndex;
+
+    if (!this.state.mounted) {
+      return false
+    }
+
+    if (this.props.rtl) {
+      nextIndex = this.state.currentSlide - this.props.slidesToScroll;
+    } else {
+      if (canGoNext(Object.assign({}, this.props,this.state))) {
+        nextIndex = this.state.currentSlide + this.props.slidesToScroll;
+      } else {
+        return false;
+      }
+    }
+
+    this.slideHandler(nextIndex);
+  };
+
+  autoPlay = () => {
+    if (this.state.autoPlayTimer) {
+      clearTimeout(this.state.autoPlayTimer);
+    }
+    if (this.props.autoplay) {
+      this.setState({
+        autoPlayTimer: setTimeout(this.play, this.props.autoplaySpeed)
+      });
+    }
+  };
+
+  pause = () => {
+    if (this.state.autoPlayTimer) {
+      clearTimeout(this.state.autoPlayTimer);
+      this.setState({
+        autoPlayTimer: null
+      });
+    }
+  };
+
+  // Event handler for previous and next
+  changeSlide = (options) => {
+    var indexOffset, previousInt, slideOffset, unevenOffset, targetSlide;
+    const {slidesToScroll, slidesToShow} = this.props
+    const {slideCount, currentSlide} = this.state
+    unevenOffset = (slideCount % slidesToScroll !== 0);
+    indexOffset = unevenOffset ? 0 : (slideCount - currentSlide) % slidesToScroll;
+
+    if (options.message === 'previous') {
+      slideOffset = (indexOffset === 0) ? slidesToScroll : slidesToShow - indexOffset;
+      targetSlide = currentSlide - slideOffset;
+      if (this.props.lazyLoad) {
+        previousInt = currentSlide - slideOffset;
+        targetSlide = previousInt === -1 ? slideCount -1 : previousInt;
+      }
+    } else if (options.message === 'next') {
+      slideOffset = (indexOffset === 0) ? slidesToScroll : indexOffset;
+      targetSlide = currentSlide + slideOffset;
+      if (this.props.lazyLoad) {
+        targetSlide = ((currentSlide + slidesToScroll) % slideCount) + indexOffset;
+      }
+    } else if (options.message === 'dots' || options.message === 'children') {
+      // Click on dots
+      targetSlide = options.index * options.slidesToScroll;
+      if (targetSlide === options.currentSlide) {
+        return;
+      }
+    } else if (options.message === 'index') {
+      targetSlide = parseInt(options.index);
+      if (targetSlide === options.currentSlide) {
+        return;
+      }
+    }
+
+    this.slideHandler(targetSlide);
+  };
+
+  // Accessiblity handler for previous and next
+  keyHandler = (e) => {
+    //Dont slide if the cursor is inside the form fields and arrow keys are pressed
+    if(!e.target.tagName.match('TEXTAREA|INPUT|SELECT')) {
+        if (e.keyCode === 37 && this.props.accessibility === true) {
+            this.changeSlide({
+              message: this.props.rtl === true ? 'next' :  'previous'
+            });
+        } else if (e.keyCode === 39 && this.props.accessibility === true) {
+            this.changeSlide({
+              message: this.props.rtl === true ? 'previous' : 'next'
+            });
+        }
+    }
+  };
+
+  // Focus on selecting a slide (click handler on track)
+  selectHandler = (options) => {
+    this.changeSlide(options)
+  };
+
+  swipeStart = (e) => {
+    var touches, posX, posY;
+
+    if ((this.props.swipe === false) || ('ontouchend' in document && this.props.swipe === false)) {
+      return;
+    } else if (this.props.draggable === false && e.type.indexOf('mouse') !== -1) {
+      return;
+    }
+    posX = (e.touches !== undefined) ? e.touches[0].pageX : e.clientX;
+    posY = (e.touches !== undefined) ? e.touches[0].pageY : e.clientY;
+    this.setState({
+      dragging: true,
+      touchObject: {
+        startX: posX,
+        startY: posY,
+        curX: posX,
+        curY: posY
+      }
+    });
+  };
+
+  swipeMove = (e) => {
+    if (!this.state.dragging) {
+      e.preventDefault();
+      return;
+    }
+    if (this.state.animating) {
+      return;
+    }
+    if (this.props.vertical && this.props.swipeToSlide && this.props.verticalSwiping) {
+      e.preventDefault();
+    }
+    var swipeLeft;
+    var curLeft, positionOffset;
+    var touchObject = this.state.touchObject;
+
+    curLeft = getTrackLeft(assign({
+      slideIndex: this.state.currentSlide,
+      trackRef: this.track
+    }, this.props, this.state));
+    touchObject.curX = (e.touches) ? e.touches[0].pageX : e.clientX;
+    touchObject.curY = (e.touches) ? e.touches[0].pageY : e.clientY;
+    touchObject.swipeLength = Math.round(Math.sqrt(Math.pow(touchObject.curX - touchObject.startX, 2)));
+
+    if (this.props.verticalSwiping) {
+      touchObject.swipeLength = Math.round(Math.sqrt(Math.pow(touchObject.curY - touchObject.startY, 2)));
+    }
+
+    positionOffset = (this.props.rtl === false ? 1 : -1) * (touchObject.curX > touchObject.startX ? 1 : -1);
+
+    if (this.props.verticalSwiping) {
+      positionOffset = touchObject.curY > touchObject.startY ? 1 : -1;
+    }
+
+    var currentSlide = this.state.currentSlide;
+    var dotCount = Math.ceil(this.state.slideCount / this.props.slidesToScroll);
+    var swipeDirection = this.swipeDirection(this.state.touchObject);
+    var touchSwipeLength = touchObject.swipeLength;
+
+    if (this.props.infinite === false) {
+      if ((currentSlide === 0 && swipeDirection === 'right') || (currentSlide + 1 >= dotCount && swipeDirection === 'left')) {
+        touchSwipeLength = touchObject.swipeLength * this.props.edgeFriction;
+
+        if (this.state.edgeDragged === false && this.props.edgeEvent) {
+          this.props.edgeEvent(swipeDirection);
+          this.setState({ edgeDragged: true });
+        }
+      }
+    }
+
+    if (this.state.swiped === false && this.props.swipeEvent) {
+      this.props.swipeEvent(swipeDirection);
+      this.setState({ swiped: true });
+    }
+
+    if (!this.props.vertical) {
+      swipeLeft = curLeft + touchSwipeLength * positionOffset;
+    } else {
+      swipeLeft = curLeft + (touchSwipeLength * (this.state.listHeight / this.state.listWidth)) * positionOffset;
+    }
+
+    if (this.props.verticalSwiping) {
+      swipeLeft = curLeft + touchSwipeLength * positionOffset;
+    }
+
+    this.setState({
+      touchObject: touchObject,
+      swipeLeft: swipeLeft,
+      trackStyle: getTrackCSS(assign({left: swipeLeft}, this.props, this.state))
+    });
+
+    if (Math.abs(touchObject.curX - touchObject.startX) < Math.abs(touchObject.curY - touchObject.startY) * 0.8)
+      { return; }
+    if (touchObject.swipeLength > 4) {
+      e.preventDefault();
+    }
+  };
+
+  getNavigableIndexes = () => {
+    let max;
+    let breakPoint = 0;
+    let counter = 0;
+    let indexes = [];
+
+    if (!this.props.infinite) {
+      max = this.state.slideCount;
+    } else {
+      breakPoint = this.props.slidesToShow * -1;
+      counter = this.props.slidesToShow * -1;
+      max = this.state.slideCount * 2;
+    }
+
+    while (breakPoint < max) {
+      indexes.push(breakPoint);
+      breakPoint = counter + this.props.slidesToScroll;
+
+      counter += this.props.slidesToScroll <= this.props.slidesToShow ?
+        this.props.slidesToScroll : this.props.slidesToShow;
+    }
+
+    return indexes;
+  };
+
+  checkNavigable = (index) => {
+    const navigables = this.getNavigableIndexes();
+    let prevNavigable = 0;
+
+    if (index > navigables[navigables.length - 1]) {
+      index = navigables[navigables.length - 1];
+    } else {
+      for (var n in navigables) {
+        if (index < navigables[n]) {
+          index = prevNavigable;
+          break;
+        }
+
+        prevNavigable = navigables[n];
+      }
+    }
+
+    return index;
+  };
+
+  getSlideCount = () => {
+    const centerOffset = this.props.centerMode ? this.state.slideWidth * Math.floor(this.props.slidesToShow / 2) : 0;
+
+    if (this.props.swipeToSlide) {
+      let swipedSlide;
+
+      const slickList = ReactDOM.findDOMNode(this.list);
+
+      const slides = slickList.querySelectorAll('.slick-slide');
+
+      Array.from(slides).every((slide) => {
+        if (!this.props.vertical) {
+          if (slide.offsetLeft - centerOffset + (this.getWidth(slide) / 2) > this.state.swipeLeft * -1) {
+            swipedSlide = slide;
+            return false;
+          }
+        } else {
+          if (slide.offsetTop + (this.getHeight(slide) / 2) > this.state.swipeLeft * -1) {
+            swipedSlide = slide;
+            return false;
+          }
+        }
+
+        return true;
+      });
+
+      const slidesTraversed = Math.abs(swipedSlide.dataset.index - this.state.currentSlide) || 1;
+
+      return slidesTraversed;
+    } else {
+      return this.props.slidesToScroll;
+    }
+  };
+
+  swipeEnd = (e) => {
+    if (!this.state.dragging) {
+      if (this.props.swipe) {
+        e.preventDefault();
+      }
+      return;
+    }
+    var touchObject = this.state.touchObject;
+    var minSwipe = this.state.listWidth/this.props.touchThreshold;
+    var swipeDirection = this.swipeDirection(touchObject);
+
+    if (this.props.verticalSwiping) {
+      minSwipe = this.state.listHeight/this.props.touchThreshold;
+    }
+
+    // reset the state of touch related state variables.
+    this.setState({
+      dragging: false,
+      edgeDragged: false,
+      swiped: false,
+      swipeLeft: null,
+      touchObject: {}
+    });
+    // Fix for #13
+    if (!touchObject.swipeLength) {
+      return;
+    }
+    if (touchObject.swipeLength > minSwipe) {
+      e.preventDefault();
+
+      let slideCount, newSlide;
+
+      switch (swipeDirection) {
+
+        case 'left':
+        case 'down':
+          newSlide = this.state.currentSlide + this.getSlideCount();
+          slideCount = this.props.swipeToSlide ? this.checkNavigable(newSlide) : newSlide;
+          this.state.currentDirection = 0;
+          break;
+
+        case 'right':
+        case 'up':
+          newSlide = this.state.currentSlide - this.getSlideCount();
+          slideCount = this.props.swipeToSlide ? this.checkNavigable(newSlide) : newSlide;
+          this.state.currentDirection = 1;
+          break;
+
+        default:
+          slideCount = this.state.currentSlide;
+
+      }
+
+      this.slideHandler(slideCount);
+    } else {
+      // Adjust the track back to it's original position.
+      var currentLeft = getTrackLeft(assign({
+        slideIndex: this.state.currentSlide,
+        trackRef: this.track
+      }, this.props, this.state));
+
+      this.setState({
+        trackStyle: getTrackAnimateCSS(assign({left: currentLeft}, this.props, this.state))
+      });
+    }
+  };
+
+  onInnerSliderEnter = (e) => {
+    if (this.props.autoplay && this.props.pauseOnHover) {
+      this.pause();
+    }
+  };
+
+  onInnerSliderOver = (e) => {
+    if (this.props.autoplay && this.props.pauseOnHover) {
+      this.pause();
+    }
+  };
+
+  onInnerSliderLeave = (e) => {
+    if (this.props.autoplay && this.props.pauseOnHover) {
+      this.autoPlay();
+    }
+  };
+}
